requirements.txt
tushare
pandas
gradio
requests
matplotlib
pyarrow
fastparquet
python-dotenv